requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.215.0
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
Pillow==10.1.0
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.4
PyPDF2==3.0.1
# Dependências para português brasileiro
spacy==3.7.2
